backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
supabase
langchain
langchain-openai
openai
sentence-transformers
mlflow
newsapi-python
python-dotenv
httpx
